requirements.txt
beautifulsoup4>=4.12.3
datasets>=2.20.0,<3.0.0
faiss-cpu>=1.8.0
fsspec>=2023.1.0,<=2024.6.1
huggingface-hub>=0.34.0,<1.0.0
numpy>=1.26.0,<2.0
openai>=1.40.0
pandas>=2.2.0
python-dotenv>=1.0.1
requests>=2.32.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0,<4.0.0
streamlit>=1.37.0
torch>=2.2.0,<2.4.0
tqdm>=4.66.0
transformers>=4.41.0,<5.0.0
